# Elixir: use built-in json + fix datetime insertion