--- a/lessons/233/elixir-app/mix.exs
+++ b/lessons/233/elixir-app/mix.exs
@@ -28,7 +28,6 @@ defmodule App.MixProject do
       {:ecto_sql, "3.12.1"},
       {:postgrex, "0.19.3"},
       {:telemetry_metrics, "1.0.0"},
-      {:jason, "1.4.4"},
       {:bandit, "1.6.1"},
       {:peep, "3.3.1"}
     ]