This pull request consists of two changes:

1. Switch from Jason to the new standard JSON library
2. Convert created_at and updated_at fields to string before insertion

### Switch from Jason to the new standard JSON library

Elixir 1.18 introduced built-in support for JSON, which according to my tests is significantly faster than Jason

```
##### With input devices_struct #####
Name                                ips        average  deviation         median         99th %
JSON.encode_to_iodata!/1       928.40 K        1.08 μs  ±4567.52%        0.93 μs        1.48 μs
Jason.encode_to_iodata!/1      494.06 K        2.02 μs  ±1756.63%        1.76 μs        3.28 μs
Comparison:
JSON.encode_to_iodata!/1       928.40 K
Jason.encode_to_iodata!/1      494.06 K - 1.88x slower +0.95 μs

##### With input devices_json #####
Name                      ips        average  deviation         median         99th %
JSON.decode!/1       328.70 K        3.04 μs   ±888.06%        2.81 μs        4.58 μs
Jason.decode!/1      235.01 K        4.26 μs   ±676.94%        3.59 μs        6.63 μs
Comparison:
JSON.decode!/1       328.70 K
Jason.decode!/1      235.01 K - 1.40x slower +1.21 μs
```

### Convert created_at and updated_at fields to string before insertion

Apparently elixir version of the app **never inserted anything into the database**.
Without `DateTime.to_string()` postgrex returns an error similar to:
`{"message":"Postgrex expected a binary, got ~U[2025-03-14 14:00:56.233724Z]. Please make sure the value you are passing matches the definition in your table or in your query or convert the value accordingly."}` with _400 Bad Request_ status code

It was present here: 6d63c9e0382f422ec6fe93d7f8e9b837ea553205
But is missing here: f991ece4c02c2963b5b9746d52d002646922d885 (the commit that you used for the benchmark)

---

\+ updated dependencies as was [recommended](https://github.com/antonputra/tutorials/pull/370#issuecomment-2561268019) by @josevalim