@josevalim 2025-03-14
Thank you @AstolfoKawaii! Good catch on the DateTime string conversion.

FWIW, I ran the benchmarks locally with and without the `to_string` conversion, and the numbers are roughly the same on my machine but maybe the servers will tell a different story. On the other hand, I'd have expected the _benchmarks to have failed_ if the server was returning 400, as some basic smoke tests should be in place.

If someone runs the benchmarks again, I'd love if the benchmarks numbers and server logs are made available for posterity. The benchmark process has been opaque: improvements we see locally are not seen in the results and the benchmarks show failures we cannot reproduce locally. Having a bit more transparency would likely help everyone: it could provide us feedback and help spot mistakes like the one in this issue.