@AstolfoKawaii 2025-03-14
> The benchmark process has been opaque: improvements we see locally are not seen in the results and the benchmarks show failures we cannot reproduce locally.

I agree. It seems like the version used for the benchmark was somehow different from the one in the repo, because the one in the repo shouldn't have worked at all 😅