@yordis 2025-03-23
We are observing huge span gaps in our traces, up to entire second or more sometimes; and most people, including myself, feels that it is due to Kubernetes CPU limiting.

Ideally the throttling would be 0 so when those high CPU usage happen it doesn't affect the system as much and let Erlang perform at 100%

@josevalim I wonder if that is the precise issue that Anton is facing in his tests. I can not find any other explanation for his and our situation by now. I used the following functions to verify, at the very least, the allocated resources:

```elixir
:erlang.system_info(:cpu_quota)
:erlang.system_info(:schedulers)
:erlang.system_info(:schedulers_online)
```

@antonputra definitely curious to see how much VMs without k8s help here, but even more curious if you remove the CPU Limiting.