@AstolfoKawaii 2025-03-23
> We are observing huge span gaps in our traces

I apologize, but could you clarify which traces you are referring to? :sweat_smile: