@antonputra 2025-04-04
@yordis Sorry for the delay. Over the last year or so, while running these benchmarks, I found that different applications, and especially configurations associated with deployment, are affected differently by Kubernetes. Once in a while, I do run my tests on VMs, but I should probably even switch to bare metal EC2s.