@yordis 2025-04-04
@antonputra we had to add Node Poll to isolate the pods, and run one pod per Node without any CPU Limits to stabilize the situation.

That could be the compromise, still k8s but treat the situation as if you are giving the entire VM to one elixir app.

The bottom line is, we can not interfere with Erlang scheduler.

Lastly, we removed the `livenessProbe` all together because it was causing issues:

> 17 Unhealthy: Liveness probe failed: command "/app/myapp/bin/myapp pid" timed out

So worth removing the `livenessProbe` and just observe the container exiting (which is the default behavior I believe)