@realhaidinh 2025-04-15
Have you tried tuning the Erlang VM? I think it might help
https://docs.pleroma.social/backend/configuration/optimizing_beam/
https://www.rabbitmq.com/docs/next/runtime#vm-settings